Take multi_edit_uploads into account (#6)
In addition to excluding the upload edit, subsequent edits by the
same user are now also excluded. This is to account for e.g. SDC
being added in separateedits despite functionally being part of the
upload.

Edits by the uploader which occur after another user has edited the
page are still included.